Initialized file format classes

+++ Pixel.h
@@ -1,7 +1,7 @@
 #pragma once
 
 struct Pixel {
-	unsigned int r, g, b;
+	size_t r, g, b;
 
-	Pixel(int _r, int _g, int _b): r(_r), g(_g), b(_b) {}
+	Pixel(size_t _r, size_t _g, size_t _b): r(_r), g(_g), b(_b) {}
 };
